Fix modules.
First remove some old code which guessed the module from the library's
key, this was there to bootstrap modules - no need now they're all
stored either in the documentation list, or implicitly by the location
of libraries.json.

Secondly, pass the name of the module as the parameter, rather than the
array containing the module data.