Normalize release data before hashing.
So that the hah values remain the same as they are on master.
Probably more effort than was required to manually fix the hashes every
time they change, but I was getting a bit fed up of that. Although I
suspect some reason will still come up so that I'll have to manually fix
the hashes again.