Remove special case for HEAD requests
It's not really any more efficient than GET, so it's mostly pointless.
I still want to go through the motions of the unzip for a HEAD request,
as I want it to fail when a GET request would fail. They're almost never
used to efficiency isn't much of a concern.