Scan for new pages when running 'update_pages --in-progress-only'
So that it will pick up new release notes. It also means that the state
will be updated for other pages. They won't be rebuilt, but I feel that
the state should only be updated when they are actually going to be
built. Maybe add an option to `update_qbk_file` to only update
development pages. Or perhaps only scan for new files that aren't in
pages yet.